+++ src/app/customize-your-seat/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/components/CustomizedSeat';
